Refactor default StorageAdapter
